SI-9834 Improve error on failed op=
If rewriting `x += y` fails to typecheck, emit error messages
for both the original tree and the assignment.

If rewrite is not attempted because `x` is a val, then say so.

The error message at `tree.pos` is updated with the additional advice.

SI-8763 Crash in update conversion

When there are already errors, don't attempt mechanical rewrites.