Use 2.12.0-M4-9901daf as STARR (see #5152)
This commit switches to using 2.12.0-M3-dc9effe as STARR,
so that we can switch to the new trait encoding where
each concrete trait member gets a a static member,
which has the actual implementation (as well as
serving as a target for for super calls using invokestatic),
and a default member (forwards to the static member).

Also bump partest to 1.0.17 -- the release that goes
with the in-sourcing of scalacheck.

Replace a few more -Yopt with -opt (for our new STARR)