SI-9847 Nuance pure expr statement warning
Clarify the current warning, which means that an expression
split over multiple lines may not be parsed as naively expected.

When typing a block, attempt minor nuance. For instance, a
single expression is not in need of parens. Try to avoid
duplicate warnings for expressions that were adapted away
from result position.