Changed From hasNext Validation to this.nonEmpty
The hasNext method actually checks to see if the next value is nonEmpty, so rather than calling ahead in the sequence through a buffered itertaing we check if this exists, otherwise we generate the None. Test case now includes last value of an Iterator, the first, and an element past the end. The nonEmpty guarantees that head will no be called on a value unless it exists, buffering the call.