renames toSet to toSetUp to highlight upcasting
toSet's behavior while sound can be surprising.
Despite Set being invariant it happily upcasts the source collection if
this makes the compiler happy, even to a collection of Any is necessary,
because the source collection is co-variant.

This compiles: `Seq(1,2,3).toSet.contains("foo")`
This does not compile: `Seq(1,2,3).toSet.map(_*2)`

Renaming to toSetUp highlights this fact.

In most cases the less surprising operation is to[Set]:
This does not compile: `Seq(1,2,3).to[Set].contains("foo")`
This compiles: `Seq(1,2,3).to[Set].map(_*2)`