Skip implicit class members in implicit search at super call
When an implicit search is triggered in a super constructor argument,
there needs to be an exception due to special scoping. Concretely, in

    object T {
      implicit val x: Int = 0
      class A(x: Int)
      class B extends A(implicitly)
    }

we get the super call

    def <init>(): T.B = B.super.<init>(implicitly)

The context nesting at the super call is something like
  - [owner=constructor B, tree=Block{B.super...}]
  - [owner=constructor B, tree=Block{B.super...}] // same owner, tree
  - [owner=constructor B, tree=Block{B.super...}] // again
  - [owner=class B, tree=ClassDef{class C ...}]
  - [owner=object T, tree=Template{extends AnyRef ...}]

When searching for the implicit, we need to skip over the context with
owner `class B`. Members (implicits) of `B` are not in scope at the
super call, by spec

  > The signature and the self constructor invocation of a constructor
  > definition are type-checked and evaluated in the scope which is in
  > effect at the point of the enclosing class definition

The `context.implicitss` implementation searches implicits in the
current context and recursively in all outer contexts. So we need to
skip the correct context in the recursive calls.

This was originally done in 198906fb11 (only 12.5 years ago...), refined
in 86a6ad44fd, refined again in 78007ac467. The current version is still
buggy when the super argument is a block (scala/bug#9717).

The `inSelfSuperCall` context flag added in 86a6ad44fd is interesting
because it's set for the context of the super call. It's used here:

    if (.. owner.isClass .. && !inSelfSuperCall)

But `implicitss` calls `outer.implicitss` recursively, and at the point
where we're `owner.isClass`, `inSelfSuperCall` is no longer defined
even if we started at the super call context. So (I think) this flag
was never doing anything.

The new implementation of this PR now checks the `inSelfSuperCall` flag.
When it's set, the enclosing class (`B` in the exmaple) is passed down
to the recursive `outer.implicitss` calls, so the class context can be
skipped. No more guessing of the context nesting structure is necessary.

Co-authored-by: Jasper Moeys <jaspermoeys@hotmail.com>
Co-authored-by: Diego Esteban Alonso Blas <diego.e.a@47deg.com>