SD-143 error for super calls that cannot be implemented correctly
If a call super[T].m resolves to a method A.m where A is a class, but
not the direct superclass of the current class, there is no way to
emit an invocation of A.m: `invokespecial A.m` will resolve to B.m
where B is the superclass of the current class.

This commit adds an error message in this case.

Note that this is similar to an existing error message for qualified
super calls to a non-direct superclass:

    class A { def m = 1 }
    class B extends A { override def m = 2 }
    class C extends B { override def m = super[A].m }

Gives "error: A does not name a parent class of class C".

If the user means to call method m in the superclass, he can write an
unqualified `super.m`.

An similar error message is introduced if A is a Java-defined interface
(and m is a default method), and A is not a direct parent of the current
class. In this case `invokespecial A.m` is invalid bytecode. The
solution is to add A as a direct parent of the current class.