Shorten fields phase description
Makes fields fit the field width, which is fitting.

`s/including/add` seems sufficient.

Possibly, "synthesize" is an extravagance for "add",
but "add" is used previously in that column.

Resolve, load, translate, add, synthesize, replace, erase,
move, eliminate, remove, generate.

Would love to learn a word that says what typer does, if
the word "type" is too redundant or overloaded, besides the
food metaphor. Also "meat-and-potatoes" implies basic,
simple, not fussy or fancy. There are many devices,
like the heart or a Ferrari engine, that are
fundamental without being unfussy.