Add clarification on preservation of order
Future.sequence does not preserve order of the original collection. The
current implementation creates a function that adds the results to the
final collection via a builder, but that function does not get called
until the Future is resolved.

So if the futures do not resolve in the
same order in which they are passed, the final collection will not have
the same order as the original.