Improve hash code of Names
The old approach of using the first, last, and middle characters
only lays a trap for generate names that have little or no entropy
at these locations.

For instance, fresh existential names generated in "as seen
from" operations are one such case, and when compiling large
batches of files the name table can become imbalanced.

This seems to be the bottleneck compiling the enourmous
(generated) test suite for ScalaTest itself:

   https://github.com/scala/scala-dev/issues/246#issuecomment-255338925

This commit uses all characters to compute the hashCode.

It improves the compilation time of ScalaTest tests from
487s to 349s (0.71x).

It would still be useful to avoid generating these fresh names
with a global counter, as this represents a steady name leak
in long-lived Globals (e.g. the presentation compiler.)