Don't follow type aliases in getClassByName and friends
This makes getClassByName fail / getClassIfDefined return NoSymbol
when querying an alias.

The current behavior can confuse the classfile parser: when parsing a
class, a cross-check verifies that `pool.getClassSymbol(nameIdx)`
returns the symbol of the class currently being parsed. If there's a
type alias that shadows the linked class, following the alias would
return an unrelated class.

(The cross-check didn't fail because there are some other guards
around it)

The logic to follow aliases was was added in ff98878, without a clear
explanation. Note that `requiredClass[C]` works if `C` is an alias, it
is expanded by the compiler.