Clean up cross-check in classfile parser, remove unnecessary assignment
One of the first entries in the classfile is the class name. The
classfile parser performs a cross-check by looking up the class sybmol
corresponding to that name and ensures it's the same as `clazz`, the
class symbol that the parser currently populates.

Since 322c980 ("Another massive IDE checkin"), if at the time of the
check `clazz` but the lookup returns some class, the `clazz` field is
assigned.

The commit following this one makes sure `clazz` is never NoSymbol, so
the assignment can safely be removed.