Silence SBT logging about macros and incremental compilation.
Since upgrading to SBT 0.13.12, clean builds have incurred warnings like:

    Because JavaMirrors.scala contains a macro definition, the following
    dependencies are invalidated unconditionally: ....

This commit disables this behaviour of the SBT incremental compiler in
the library and reflect projects, as these aren't regular macros (the
macro implementations are hard coded in the compiler in `FastTrack`)
so the new behaviour isn't actually improving correctness of inc.
compilation.