SI-9814 Fix synchronized in specialized overrides
Specialization creates a subclasses of a specializd class for each
type parameter combination. These contains copies of the methods from
the superclass.

However, before this transform, the pattern of self-synchronization
in a method body had been replace by flag Flag.SYNCHRONIZED on the
method symbol. This was not being propagated to the override, and
hence no locking occured.

This commit modifies the creation of the specialized overload symbol
to copy the SYNCHRONIZED flag, as was already done for ASBOVERRIDE.

I have also done the same for the `@strictfp` annotation.